+++ server/component/clockwork/ClockworkView.php
@@ -1,0 +1,53 @@
+<?php
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+?>
+<?php
+require_once __DIR__ . "/../BaseView.php";
+
+/**
+ * The view class of the asset select component.
+ */
+class ClockworkView extends BaseView
+{
+
+    /* Private Properties *****************************************************/
+
+    /* Constructors ***********************************************************/
+
+    /**
+     * The constructor.
+     *
+     * @param object $model
+     *  The model instance of the component.
+     * @param object $controller
+     *  The controller instance of the component.
+     */
+    public function __construct($model, $controller)
+    {
+        parent::__construct($model, $controller);
+    }
+
+    /* Private Methods ********************************************************/
+
+    /* Public Methods *********************************************************/
+
+    /**
+     * Render the view.
+     */
+    public function output_content()
+    {
+        if ($this->model->is_clockwork_enabled()) {
+            $this->output_controller_alerts_success(true);
+        }else{
+            $this->output_controller_alerts_fail(true);
+        }
+    }
+
+    public function output_content_mobile()
+    {
+        return;
+    }
+}
+?>