+++ server/db/update_scripts/34_update_v7.2.0_v7.3.0.sql
@@ -1,0 +1,27 @@
+-- set DB version
+UPDATE version
+SET version = 'v7.3.0';
+
+-- add clockwork page
+INSERT IGNORE INTO `pages` (`id`, `keyword`, `url`, `protocol`, `id_actions`, `id_navigation_section`, `parent`, `is_headless`, `nav_position`, `footer_position`, `id_type`, `id_pageAccessTypes`) 
+VALUES (NULL, 'clockwork', '/admin/clockwork', 'GET|POST', '0000000002', NULL, '0000000009', '0', '900', NULL, '0000000001', (SELECT id FROM lookups WHERE type_code = "pageAccessTypes" AND lookup_code = "web"));
+
+SET @id_page = (SELECT id FROM pages WHERE keyword = 'clockwork');
+
+INSERT IGNORE INTO `pages_fields_translation` (`id_pages`, `id_fields`, `id_languages`, `content`) VALUES (@id_page, get_field_id('label'), '0000000002', 'Clockwork');
+INSERT IGNORE INTO `pages_fields_translation` (`id_pages`, `id_fields`, `id_languages`, `content`) VALUES (@id_page, get_field_id('title'), '0000000002', 'Clockwork');
+INSERT IGNORE INTO `acl_groups` (`id_groups`, `id_pages`, `acl_select`, `acl_insert`, `acl_update`, `acl_delete`) VALUES ('0000000001', @id_page, '1', '0', '0', '0');
+
+-- add column clockwork to table cmsPreferences 
+CALL add_table_column('cmsPreferences', 'clockwork', "INT(11) DEFAULT '0'");
+
+DROP VIEW IF EXISTS view_cmsPreferences;
+CREATE VIEW view_cmsPreferences
+AS
+SELECT p.callback_api_key, p.default_language_id, l.`language` AS default_language, l.locale, p.firebase_config, p.anonymous_users, p.clockwork
+FROM cmsPreferences p
+LEFT JOIN languages l ON (l.id = p.default_language_id)
+WHERE p.id = 1;
+
+
+